refactor: AppContainerStore에서 AppContainer에 등록한 인터페이스 처리 가능하도록 변경
