+++ app/src/main/java/woowacourse/shopping/ui/MainApplication.kt
@@ -2,7 +2,9 @@ package woowacourse.shopping.ui
 
 import android.app.Application
 import woowacourse.shopping.di.AppContainer
+import woowacourse.shopping.di.AppContainerStore
 
 class MainApplication : Application() {
     val appContainer = AppContainer()
+    val appContainerStore = AppContainerStore(appContainer)
 }
